+++ test/loader.js
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2018 Rafael da Silva Rocha.
+ * Copyright (c) 2018-2019 Rafael da Silva Rocha.
  * https://github.com/rochars/utf8-buffer
  *
  */
@@ -9,13 +9,6 @@ let utf8Buffer;
 // UMD bundle
 if (process.argv[3] == '--umd') {
 	console.log('umd tests');
-	utf8Buffer = require('../dist/utf8-buffer.umd.js');
-
-// ES6 dist
-} else if (process.argv[3] == '--esm') {
-	require = require("esm")(module);
-	global.module = module;
-	console.log("esm");
 	utf8Buffer = require('../dist/utf8-buffer.js');
 
 // Source
@@ -23,7 +16,7 @@ if (process.argv[3] == '--umd') {
 	require = require("esm")(module);
 	global.module = module;
 	console.log('Source tests');
-	utf8Buffer = require('../main.js');
+	utf8Buffer = require('../index.js');
 }
 
 module.exports = utf8Buffer;